chat/requirements.txt
streamlit==1.28.0
openai==1.3.0
langchain==0.0.339
langchain-core>=0.1.0
langchain-community==0.0.6
python-dotenv==1.0.0
requests==2.31.0
neo4j==5.14.1
pandas==2.1.3
plotly==5.18.0
# Phoenix / Agents
openinference-instrumentation-openai-agents>=0.0.1
openai-agents>=0.0.1
arize-phoenix>=1.0.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp>=1.21.0
google-generativeai # If needed later, keeping common deps
